refactor: 移除 `Equipment` 分類
- 解耦 `Entity` -> `Equipment` ，讓 `GunEntity` 直接繼承 `Entity`
- 取消共通型別 `GunEntity`，針對每個實體建立獨立的物件
- 將 `ComponentDataType['xigmaguns:timer']` 更改為陣列形式
