--- src/declare/entity/equipment/Equipment.d.ts
@@ -1,8 +1,0 @@
-declare type EquipmentType = keyof Pick<EntityTypes, 'Gun' | 'Melee' | 'Throwable'>;
-
-declare interface GunTypes {
-    Glock17: 'Glock17',
-    Unknown: 'Unknown',
-}
-
-declare type GunType = keyof GunTypes;
