+++ src/source/entity/Entity.ts
@@ -2,15 +2,18 @@ import { ComponentTypes, Component } from "../components/Component";
 
 export interface IEntity {
     readonly uuid: string;
-    readonly components: Map<keyof ComponentTypes, Component | Component[]>
+    readonly type: EntityType;
+    readonly components: Map<keyof ComponentTypes, Component | Component[]>;
 }
 
 export class Entity implements IEntity {
 
     readonly uuid: string;
+    readonly type: EntityType;
     readonly components = new Map;
 
-    constructor(uuid: string) {
+    constructor(uuid: string, type: EntityType) {
         this.uuid = uuid;
+        this.type = type;
     }
 }
